// src/app/models/grupo.ts
export class Grupo {
    segmento:            string;
    codigo:              string;
    nome:                string;
    nivel:               string;
    grupo:               string;
    comandoGrupo:        string;
    email:               string;

    // adicionais
    perfil: string;

    constructor() {
        this.segmento = '';
        this.codigo = '';
        this.nome = '';
        this.nivel = '';
        this.grupo = '';
        this.comandoGrupo = '';
        this.email = '';
    }
}